+++ docs/changelog.rst
@@ -5,6 +5,7 @@
 ----------
 - :attr:`scapi.User.get_activities` を追加
 - :attr:`scapi.Classroom.get_public_activity` を追加
+- :attr:`scapi.Session.get_message_count` が正しい値を取得しない問題を修正
 
 3.x.x
 -----